Move in-model CP layout helpers out of training_utils into model plugins
miles/backends/training_utils/cp_utils.py owns the CP helpers the training
backend applies around the model: slicing data, masks, logprobs and logits.
build_gdn_cp_context was the one exception -- it runs inside a layer forward,
is used only by model plugins, and forced an optional flash-linear-attention
dependency (the fla.ops.cp try/except) into a core training module.

The zigzag <-> contiguous-shard relayout has the same shape and was private to
miles_plugins/models/hf_attention.py, so only the two models built on
HuggingfaceAttention could reach it.

Collect both in miles_plugins/models/cp_utils.py, which now holds the CP
helpers that run inside a forward. To break the resulting import cycle,
detect_and_setup_hybrid_cp moves the other way, into hf_attention.py next to
the HuggingfaceAttention class it is defined in terms of.

Pure move: every relocated definition is byte-identical apart from dropping the
underscore prefix on the four newly public relayout helpers.
