+++ miles_plugins/models/qwen3_next.py
@@ -16,7 +16,7 @@ try:
 except ImportError:
     pass
 
-from miles.backends.training_utils.cp_utils import build_gdn_cp_context
+from miles_plugins.models.cp_utils import build_gdn_cp_context
 
 from .hf_attention import HuggingfaceAttention
 from .qwen_gdn_backend import get_chunk_gated_delta_rule
